refactor(structure): phase 1 — drop unused hook, promote callingLabels to lib
- Delete src/hooks/useCommentUnread.ts (no importers).
- Promote src/features/settings/callingLabels.ts → src/lib/callingLabels.ts
  (cross-feature: settings, invites, templates, app/routes import it).
- Update 5 importers to @/lib/callingLabels.

First slice of the per-component-barrel refactor. Behavior unchanged;
typecheck + 346 tests pass.
