+++ src/features/invites/PendingInvitesList.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { cn } from "@/lib/cn";
 import { friendlyWriteError } from "@/stores/saveStatusStore";
-import { CALLING_LABELS } from "../settings/callingLabels";
+import { CALLING_LABELS } from "@/lib/callingLabels";
 import { revokeInvite } from "./inviteActions";
 import { useWardInvites } from "./useWardInvites";
