+++ src/features/settings/MemberRow.tsx
@@ -2,7 +2,7 @@ import { Avatar } from "@/components/ui/Avatar";
 import { Select } from "@/components/ui/Select";
 import type { WithId } from "@/hooks/_sub";
 import type { Calling, Member } from "@/lib/types";
-import { CALLING_OPTIONS } from "./callingLabels";
+import { CALLING_OPTIONS } from "@/lib/callingLabels";
 import { LastBishopricError, setActive, setCcOnEmails, updateCalling } from "./memberActions";
 
 interface Props {
